guide link

+++ game.js
@@ -112,6 +112,8 @@ function returningUser(back) {
 	const titleText = back === "noback" ? `Welcome, ${username}!` : `Welcome back, ${username}!`;
 	document.getElementById('ruModal-content').innerHTML = `
     <h3 class="font-bold text-lg">${titleText}</h3>
+	<h1 class="text-md"><i class="fas fa-book mr-2 fa-xl"></i>Need help? Check out the <a target="_blank" class='underline' href=" https:bean-frog.github.io/blog/index.html?articleId=4">Guide</a></h1>
+
     <div class="modal-action">
       <form method="dialog">
         <button onclick='begingame();' class="btn">Start Game</button>
@@ -122,10 +124,14 @@ function returningUser(back) {
 }
 
 function wipeStats() {
-	localStorage.setItem('tuneguessr-incorrect', '');
+	if (window.confirm("This action is irreversable! Do you want to continue?")) {
+		localStorage.setItem('tuneguessr-incorrect', '');
 	localStorage.setItem('tuneguessr-correct', '');
 	localStorage.setItem('tuneguessr-total', '');
 	updateStats()
+	} else {
+		alert("canceled, be careful next time");
+	}
 }
 
 function changeUsername(newName) {
